# api/openapi.yaml
openapi: "3.1.0"
info:
  title: Receipt Parser API
  version: "1.0.0"
  description: |
    API for asynchronously or synchronously parsing receipt images. Upload an image (multipart),
    optionally provide structured metadata as a JSON part, poll job status, and fetch results.
servers:
  - url: "https://api.example.com/v1"
    description: Production server
  - url: "http://localhost:8000/v1"
    description: Local development
tags:
  - name: receipts
    description: Receipt parsing operations
paths:
  /receipts:
    post:
      tags: [receipts]
      summary: Upload a receipt image for parsing
      description: |
        Accepts an image file (multipart) and returns a job id for asynchronous processing. Supports
        an optional query parameter `sync=true` which will attempt a synchronous parse and return
        the parse result directly when possible.
      operationId: uploadReceipt
      parameters:
        - name: sync
          in: query
          schema:
            type: boolean
          description: If true, attempt synchronous parsing (may return 200 or 202)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                metadata:
                  $ref: '#/components/schemas/Metadata'
            encoding:
              file:
                contentType: "image/png, image/jpeg, image/tiff"
              metadata:
                contentType: application/json
      responses:
        "202":
          description: Accepted - job queued
          headers:
            Location:
              $ref: '#/components/headers/Location'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadAccepted'
              examples:
                uploadAccepted:
                  $ref: '#/components/examples/UploadAcceptedExample'
        "200":
          description: Synchronous parse result (if sync=true)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParseResult'
              examples:
                parseResult:
                  $ref: '#/components/examples/ParseResultExample'
        "400":
          $ref: '#/components/responses/BadRequest'
        "422":
          $ref: '#/components/responses/Unparsable'
  /receipts/parse:
    post:
      tags: [receipts]
      summary: Upload a receipt image for synchronous parsing
      description: |
        Accepts an image file (multipart) and returns the parse result directly.
      operationId: parseReceiptSync
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
            encoding:
              file:
                contentType: "image/png, image/jpeg, image/tiff"
      responses:
        "200":
          description: Synchronous parse result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParseResult'
              examples:
                parseResult:
                  $ref: '#/components/examples/ParseResultExample'
        "400":
          $ref: '#/components/responses/BadRequest'
        "422":
          $ref: '#/components/responses/Unparsable'
  /receipts/{job_id}/status:
    get:
      tags: [receipts]
      summary: Get job status
      operationId: getJobStatus
      parameters:
        - $ref: '#/components/parameters/job_id'
      responses:
        "200":
          description: Job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatus'
        "404":
          $ref: '#/components/responses/NotFound'
  /receipts/{job_id}:
    get:
      tags: [receipts]
      summary: Get parse result
      operationId: getParseResult
      parameters:
        - $ref: '#/components/parameters/job_id'
      responses:
        "200":
          description: Parse result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParseResult'
        "202":
          description: Still processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatus'
        "404":
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  parameters:
    job_id:
      name: job_id
      in: path
      required: true
      schema:
        type: string
      description: The job identifier returned by the upload endpoint
  headers:
    Location:
      description: URL to poll for job status and result
      schema:
        type: string
        format: uri
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            notFoundError:
              $ref: '#/components/examples/ErrorNotFoundExample'
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            badRequestError:
              $ref: '#/components/examples/ErrorBadRequestExample'
    Unparsable:
      description: Uploaded data could not be parsed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            unparsableError:
              $ref: '#/components/examples/ErrorUnparsableExample'
  schemas:
    Metadata:
      type: object
      description: Optional structured metadata about the upload. Sent as a JSON part in the multipart request.
      properties:
        locale:
          type: string
          description: BCP 47 language tag (e.g. 'en-US')
          example: en-US
        currency:
          type: string
          description: 3-letter ISO currency code (optional, used to hint parser)
          minLength: 3
          maxLength: 3
          example: USD
        request_id:
          type: string
          description: Client-provided id for correlation (optional)
        merchant_hint:
          type: string
          description: Optional merchant name hint to help parsing
        timezone:
          type: string
          description: IANA timezone name (e.g. 'Europe/Berlin')
        source:
          type: string
          description: Source of the image (mobile, email, upload-service)
      additionalProperties: false
    UploadAccepted:
      type: object
      required: [job_id, status_url]
      properties:
        job_id:
          type: string
        status_url:
          type: string
          format: uri
        estimated_seconds:
          type: integer
          minimum: 0
    JobStatus:
      type: object
      required: [job_id, status]
      properties:
        job_id:
          type: string
        status:
          type: string
          enum: [queued, processing, completed, failed]
        progress:
          type: integer
          minimum: 0
          maximum: 100
          description: Percentage 0-100
        error:
          $ref: '#/components/schemas/Error'
    ParseResult:
      type: object
      required: [job_id, status, parsed]
      properties:
        job_id:
          type: string
        status:
          type: string
          enum: [completed]
        parsed:
          type: object
          required: [merchant, date, total]
          properties:
            merchant:
              type: string
            date:
              type: string
              format: date
            currency:
              type: string
              example: USD
            total:
              type: number
              minimum: 0
            subtotal:
              type: number
              minimum: 0
            tax:
              type: number
              minimum: 0
            line_items:
              type: array
              items:
                $ref: '#/components/schemas/LineItem'
            raw_text:
              type: string
            warnings:
              type: array
              items:
                type: string
        meta:
          type: object
          properties:
            processing_time_seconds:
              type: number
              minimum: 0
            model_version:
              type: string
    LineItem:
      type: object
      required: [description, total_price]
      properties:
        description:
          type: string
        qty:
          type: number
          minimum: 0
        unit_price:
          type: number
          minimum: 0
        total_price:
          type: number
          minimum: 0
        confidence:
          type: number
          minimum: 0
          maximum: 1
        bounding_box:
          type: array
          description: [x, y, width, height]
          items:
            type: integer
            minimum: 0
          minItems: 4
          maxItems: 4
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            oneOf:
              - type: string
              - type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
  examples:
    UploadAcceptedExample:
      summary: Example response when job is queued
      value:
        job_id: "abc123"
        status_url: "https://api.example.com/v1/receipts/abc123"
        estimated_seconds: 12
    ParseResultExample:
      summary: Example successful parse result
      value:
        job_id: "abc123"
        status: "completed"
        parsed:
          merchant: "Acme Store"
          date: "2025-11-01"
          currency: "USD"
          total: 23.45
          subtotal: 20.00
          tax: 3.45
          line_items:
            - description: "1x Coffee"
              qty: 1
              unit_price: 3.50
              total_price: 3.50
              confidence: 0.98
              bounding_box: [10, 20, 100, 20]
          raw_text: "Acme Store\n1x Coffee 3.50\nTotal 23.45"
          warnings: []
        meta:
          processing_time_seconds: 1.23
          model_version: "v1.2.3"
    ErrorBadRequestExample:
      summary: Bad request example
      value:
        code: "BAD_REQUEST"
        message: "Invalid multipart body: 'file' part is required"
        details:
          - "multipart: missing 'file' part"
    ErrorUnparsableExample:
      summary: Unparsable upload example
      value:
        code: "UNPARSABLE"
        message: "Uploaded data could not be parsed"
        details:
          - field: "image"
            message: "No text detected in image"
    ErrorNotFoundExample:
      summary: Not found example
      value:
        code: "NOT_FOUND"
        message: "Job not found"
        details:
          - "No job with the provided id"
security:
  - BearerAuth: []